+++ main.py
@@ -1,47 +1,12 @@
 #!/usr/bin/env python3
-
-import json
 import logging
 import readline
-import signal
 import sys
 
+from twisted.internet import reactor, stdio
+
 from core import Config, Kapellmeister
-from utils import timeout
-
-
-_LOGGER = logging.getLogger(__name__)
-
-
-def _await_reply(channel):
-    result = b''
-    with timeout(1):
-        while not result:
-            result = channel.read()
-    return result
-
-
-# FIXME
-def stop_all(km):
-    _LOGGER.warn("stop all")
-
-
-def mainloop(km):
-    channel = km.connect("brain")
-
-    while True:
-        s = input("> ")
-        channel.write(json.dumps({"message": s,
-                                  "from": {"user": "user",
-                                           "channel": "channel"},
-                                  "to": {"user": "niege",
-                                         "channel": "brain"}}).encode())
-        result = _await_reply(channel).decode()
-        print("Niege>", json.loads(result)['message'])
-
-
-def _term(*_):
-    exit()
+from server import ChatProtocol
 
 
 def main():
@@ -49,15 +14,13 @@ def main():
         config = Config(cfg.read())
     km = Kapellmeister(config)
     km.run()
+    channel = km.connect("brain")
 
-    try:
-        mainloop(km)
-    except KeyboardInterrupt:
-        _LOGGER.info("Exiting on keyboard interrupt")
+    stdio.StandardIO(ChatProtocol(channel))
+    reactor.run()
 
 
 if __name__ == '__main__':
-    signal.signal(signal.SIGTERM, _term)
     logging.basicConfig(stream=sys.stderr, level=logging.DEBUG)
     readline.set_auto_history(True)